combine formerly separate metrics and diagnostics computations
