+++ workflows/diagnostics/fv3net/diagnostics/offline/compute_diagnostics.py
@@ -1,6 +1,6 @@
 from fv3net.diagnostics._shared.registry import Registry, prepare_diag_dict
 import fv3net.diagnostics._shared.transform as transform
-from fv3net.diagnostics._shared.constants import DiagArg, SURFACE_TYPE
+from fv3net.diagnostics._shared.constants import DiagArg, HORIZONTAL_DIMS, SURFACE_TYPE
 import logging
 import numpy as np
 import pandas as pd
@@ -8,7 +8,6 @@ from typing import Sequence, Tuple, Dict, Union, Mapping, Optional
 import xarray as xr
 
 from vcm import thermo, safe, weighted_average, local_time
-
 
 logger = logging.getLogger(__name__)
 
@@ -41,8 +40,9 @@ def compute_diagnostics(
     delp: xr.DataArray,
     n_jobs: int = -1,
 ):
-    diag_arg = DiagArg(prediction, target, grid, delp=delp)
-    return diagnostics_registry.compute(diag_arg, n_jobs=n_jobs)
+    return diagnostics_registry.compute(
+        DiagArg(prediction, target, grid, delp), n_jobs=n_jobs
+    )
 
 
 def conditional_average_over_domain(
@@ -201,6 +201,228 @@ def _predicts_sphum_tendency(ds):
         if "q2" in var.lower():
             return True
     return False
+
+
+def mse(x, y, w, dims):
+    with xr.set_options(keep_attrs=True):
+        return ((x - y) ** 2 * w).sum(dims) / w.sum(dims)
+
+
+def bias(truth, prediction, w, dims):
+    with xr.set_options(keep_attrs=True):
+        return prediction - truth
+
+
+def weighted_mean(ds, weights, dims):
+    with xr.set_options(keep_attrs=True):
+        return (ds * weights).sum(dims) / weights.sum(dims)
+
+
+def zonal_mean(
+    ds: xr.Dataset, latitude: xr.DataArray, bins=np.arange(-90, 91, 2)
+) -> xr.Dataset:
+    with xr.set_options(keep_attrs=True):
+        zm = (
+            ds.groupby_bins(latitude, bins=bins)
+            .mean(skipna=True)
+            .rename(lat_bins="latitude")
+        )
+    latitude_midpoints = [x.item().mid for x in zm["latitude"]]
+    return zm.assign_coords(latitude=latitude_midpoints)
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"mse_2d_{mask_type}")
+    @transform.apply(transform.select_2d_variables)
+    @transform.apply(transform.mask_area, mask_type)
+    def mse_2d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing mean squared errors for 2D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        mse_area_weighted_avg = weighted_mean(
+            (predicted - target) ** 2, weights=grid.area, dims=HORIZONTAL_DIMS
+        )
+        return mse_area_weighted_avg.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"mse_pressure_level_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_area, mask_type)
+    def mse_3d(diag_arg, mask_type=mask_type):
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        logger.info(f"Preparing mean squared errors for 3D variables, {mask_type}")
+        if len(predicted) == 0:
+            return xr.Dataset()
+        mse_area_weighted_avg = weighted_mean(
+            (predicted - target) ** 2, weights=grid.area, dims=HORIZONTAL_DIMS
+        )
+        return mse_area_weighted_avg.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"variance_2d_{mask_type}")
+    @transform.apply(transform.select_2d_variables)
+    @transform.apply(transform.mask_area, mask_type)
+    def variance_2d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing variance for 2D variables, {mask_type}")
+        target, grid = diag_arg.verification, diag_arg.grid
+        mean = weighted_mean(target, weights=grid.area, dims=HORIZONTAL_DIMS).mean(
+            "time"
+        )
+        return weighted_mean(
+            (mean - target) ** 2, weights=grid.area, dims=HORIZONTAL_DIMS
+        ).mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"variance_pressure_level_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_area, mask_type)
+    def variance_3d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing variance for 3D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        if len(predicted) == 0:
+            return xr.Dataset()
+        mean = weighted_mean(target, weights=grid.area, dims=HORIZONTAL_DIMS).mean(
+            "time"
+        )
+        return weighted_mean(
+            (mean - target) ** 2, weights=grid.area, dims=HORIZONTAL_DIMS
+        ).mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"bias_2d_{mask_type}")
+    @transform.apply(transform.select_2d_variables)
+    @transform.apply(transform.mask_area, mask_type)
+    def bias_2d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing biases for 2D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        biases_area_weighted_avg = weighted_mean(
+            predicted - target, weights=grid.area, dims=HORIZONTAL_DIMS
+        )
+        return biases_area_weighted_avg.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"bias_pressure_level_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_area, mask_type)
+    def bias_3d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing biases for 3D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        if len(predicted) == 0:
+            return xr.Dataset()
+        biases_area_weighted_avg = weighted_mean(
+            predicted - target, weights=grid.area, dims=HORIZONTAL_DIMS
+        )
+        return biases_area_weighted_avg.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"bias_2d_zonal_avg_{mask_type}")
+    @transform.apply(transform.select_2d_variables)
+    @transform.apply(transform.mask_to_sfc_type, mask_type)
+    def bias_zonal_avg_2d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing zonal avg biases for 2D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        zonal_avg_bias = zonal_mean(predicted - target, grid.lat)
+        return zonal_avg_bias.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"bias_pressure_level_zonal_avg_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_to_sfc_type, mask_type)
+    def bias_zonal_avg_3d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing zonal avg biases for 3D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        if len(predicted) == 0:
+            return xr.Dataset()
+        zonal_avg_bias = zonal_mean(predicted - target, grid.lat)
+        return zonal_avg_bias.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"mse_pressure_level_zonal_avg_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_to_sfc_type, mask_type)
+    def mse_zonal_avg_3d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing zonal avg MSEs for 3D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        if len(predicted) == 0:
+            return xr.Dataset()
+        zonal_avg_mse = zonal_mean((predicted - target) ** 2, grid.lat)
+        return zonal_avg_mse.mean("time")
+
+
+for mask_type in ["global", "sea", "land"]:
+
+    @diagnostics_registry.register(f"variance_pressure_level_zonal_avg_{mask_type}")
+    @transform.apply(transform.select_3d_variables)
+    @transform.apply(transform.regrid_zdim_to_pressure_levels)
+    @transform.apply(transform.mask_to_sfc_type, mask_type)
+    @transform.apply(transform.mask_area, mask_type)
+    def variance_zonal_avg_3d(diag_arg, mask_type=mask_type):
+        logger.info(f"Preparing zonal avg variance for 3D variables, {mask_type}")
+        predicted, target, grid = (
+            diag_arg.prediction,
+            diag_arg.verification,
+            diag_arg.grid,
+        )
+        if len(predicted) == 0:
+            return xr.Dataset()
+        mean = weighted_mean(target, weights=grid.area, dims=HORIZONTAL_DIMS).mean(
+            "time"
+        )
+        zonal_avg_variance = zonal_mean((mean - target) ** 2, grid.lat)
+        return zonal_avg_variance.mean("time")
 
 
 for mask_type in ["global", "land", "sea"]:
